fixed using comments

+++ Objective-C_HomeWorks/HW05Arrays/HW05Arrays/EMAnimal.h
@@ -10,9 +10,9 @@
 
 @interface EMAnimal : NSObject
 
-@property (strong, nonatomic) NSString* type;
-@property (strong, nonatomic) NSString* nickname;
-@property (assign, nonatomic) UInt8 year;
+@property (strong, nonatomic) NSString *type;
+@property (strong, nonatomic) NSString *nickname;
+@property (assign, nonatomic) NSInteger year;
 
 - (void) moving;
